Move more Assembly methods to shared partition
